Move ForbiddenCode-related tests to the corresponding test file
