+++ tests/ForbiddenCodeTest.php
@@ -1,5 +1,13 @@
 <?php
+
 namespace Psalm\Tests;
+
+use Psalm\Config;
+use Psalm\Context;
+use Psalm\Tests\Internal\Provider;
+
+use function dirname;
+use function getcwd;
 
 class ForbiddenCodeTest extends TestCase
 {
@@ -61,5 +69,216 @@ class ForbiddenCodeTest extends TestCase
                     }',
             ],
         ];
+    }
+
+    public function testAllowedEchoFunction(): void
+    {
+        $this->project_analyzer = $this->getProjectAnalyzerWithConfig(
+            TestConfig::loadFromXML(
+                dirname(__DIR__, 2),
+                '<?xml version="1.0"?>
+                <psalm></psalm>'
+            )
+        );
+
+        $file_path = getcwd() . '/src/somefile.php';
+
+        $this->addFile(
+            $file_path,
+            '<?php
+                echo "hello";'
+        );
+
+        $this->analyzeFile($file_path, new Context());
+    }
+
+    public function testForbiddenEchoFunctionViaFunctions(): void
+    {
+        $this->expectExceptionMessage('ForbiddenCode');
+        $this->expectException(\Psalm\Exception\CodeException::class);
+        $this->project_analyzer = $this->getProjectAnalyzerWithConfig(
+            TestConfig::loadFromXML(
+                dirname(__DIR__, 2),
+                '<?xml version="1.0"?>
+                <psalm>
+                    <forbiddenFunctions>
+                        <function name="echo" />
+                    </forbiddenFunctions>
+                </psalm>'
+            )
+        );
+
+        $file_path = getcwd() . '/src/somefile.php';
+
+        $this->addFile(
+            $file_path,
+            '<?php
+                echo "hello";'
+        );
+
+        $this->analyzeFile($file_path, new Context());
+    }
+
+    public function testForbiddenEchoFunctionViaFlag(): void
+    {
+        $this->expectExceptionMessage('ForbiddenEcho');
+        $this->expectException(\Psalm\Exception\CodeException::class);
+        $this->project_analyzer = $this->getProjectAnalyzerWithConfig(
+            TestConfig::loadFromXML(
+                dirname(__DIR__, 2),
+                '<?xml version="1.0"?>
+                <psalm forbidEcho="true"></psalm>'
+            )
+        );
+
+        $file_path = getcwd() . '/src/somefile.php';
+
+        $this->addFile(
+            $file_path,
+            '<?php
+                echo "hello";'
+        );
+
+        $this->analyzeFile($file_path, new Context());
+    }
+
+    public function testAllowedPrintFunction(): void
+    {
+        $this->project_analyzer = $this->getProjectAnalyzerWithConfig(
+            TestConfig::loadFromXML(
+                dirname(__DIR__, 2),
+                '<?xml version="1.0"?>
+                <psalm></psalm>'
+            )
+        );
+
+        $file_path = getcwd() . '/src/somefile.php';
+
+        $this->addFile(
+            $file_path,
+            '<?php
+                print "hello";'
+        );
+
+        $this->analyzeFile($file_path, new Context());
+    }
+
+    public function testForbiddenPrintFunction(): void
+    {
+        $this->expectExceptionMessage('ForbiddenCode');
+        $this->expectException(\Psalm\Exception\CodeException::class);
+        $this->project_analyzer = $this->getProjectAnalyzerWithConfig(
+            TestConfig::loadFromXML(
+                dirname(__DIR__, 2),
+                '<?xml version="1.0"?>
+                <psalm>
+                    <forbiddenFunctions>
+                        <function name="print" />
+                    </forbiddenFunctions>
+                </psalm>'
+            )
+        );
+
+        $file_path = getcwd() . '/src/somefile.php';
+
+        $this->addFile(
+            $file_path,
+            '<?php
+                print "hello";'
+        );
+
+        $this->analyzeFile($file_path, new Context());
+    }
+
+    public function testAllowedVarExportFunction(): void
+    {
+        $this->project_analyzer = $this->getProjectAnalyzerWithConfig(
+            TestConfig::loadFromXML(
+                dirname(__DIR__, 2),
+                '<?xml version="1.0"?>
+                <psalm></psalm>'
+            )
+        );
+
+        $file_path = getcwd() . '/src/somefile.php';
+
+        $this->addFile(
+            $file_path,
+            '<?php
+                $a = [1, 2, 3];
+                var_export($a);'
+        );
+
+        $this->analyzeFile($file_path, new Context());
+    }
+
+    public function testForbiddenVarExportFunction(): void
+    {
+        $this->expectExceptionMessage('ForbiddenCode');
+        $this->expectException(\Psalm\Exception\CodeException::class);
+        $this->project_analyzer = $this->getProjectAnalyzerWithConfig(
+            TestConfig::loadFromXML(
+                dirname(__DIR__, 2),
+                '<?xml version="1.0"?>
+                <psalm>
+                    <forbiddenFunctions>
+                        <function name="var_export" />
+                    </forbiddenFunctions>
+                </psalm>'
+            )
+        );
+
+        $file_path = getcwd() . '/src/somefile.php';
+
+        $this->addFile(
+            $file_path,
+            '<?php
+                $a = [1, 2, 3];
+                var_export($a);'
+        );
+
+        $this->analyzeFile($file_path, new Context());
+    }
+
+    public function testForbiddenEmptyFunction(): void
+    {
+        $this->expectExceptionMessage('ForbiddenCode');
+        $this->expectException(\Psalm\Exception\CodeException::class);
+        $this->project_analyzer = $this->getProjectAnalyzerWithConfig(
+            TestConfig::loadFromXML(
+                dirname(__DIR__, 2),
+                '<?xml version="1.0"?>
+                <psalm>
+                    <forbiddenFunctions>
+                        <function name="empty" />
+                    </forbiddenFunctions>
+                </psalm>'
+            )
+        );
+
+        $file_path = getcwd() . '/src/somefile.php';
+
+        $this->addFile(
+            $file_path,
+            '<?php
+                empty(false);'
+        );
+
+        $this->analyzeFile($file_path, new Context());
+    }
+
+    private function getProjectAnalyzerWithConfig(Config $config): \Psalm\Internal\Analyzer\ProjectAnalyzer
+    {
+        $p = new \Psalm\Internal\Analyzer\ProjectAnalyzer(
+            $config,
+            new \Psalm\Internal\Provider\Providers(
+                $this->file_provider,
+                new Provider\FakeParserCacheProvider()
+            )
+        );
+
+        $p->setPhpVersion('7.4');
+
+        return $p;
     }
 }
